0518 papago api 선언 위치 변경
